pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
name = "spVIPESmulti"
version = "1.0.0"
description = "Shared-private Variational Inference with Product of Experts and Supervision"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Mikhael Dito Manurung"},
    {name = "Claudio Novella Rausell"},
]
maintainers = [
    {name = "Mikhael Dito Manurung"},
    {name = "Claudio Novella Rausell", email = "c.novella_rausell@lumc.nl"},
]
urls.Documentation = "https://spVIPESmulti.readthedocs.io/"
urls.Source = "https://github.com/mdmanurung/spVIPESmulti"
urls.Home-page = "https://github.com/mdmanurung/spVIPESmulti"
dependencies = [
    "anndata>=0.10",
    # for debug logging (referenced from the issue template)
    "session-info",
    "scvi-tools>=1.0,<2",
    "torch",
    "zuko>=1.0.0",
    "scanpy>=1.9.1",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "twine>=4.0.2"
]
doc = [
    "docutils>=0.8,!=0.18.*,!=0.19.*",
    "sphinx>=4",
    "sphinx-book-theme>=1.0.0",
    "myst-nb",
    "sphinxcontrib-bibtex>=1.0.0",
    "sphinx-autodoc-typehints",
    "sphinxext-opengraph",
    # For notebooks
    "ipykernel",
    "ipython",
    "sphinx-copybutton",
]
test = [
    "pytest",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ['src/spVIPESmulti']

[tool.coverage.run]
source = ["spVIPESmulti"]
omit = [
    "**/test_*.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",  # allow using test files with same name
]
markers = [
    "integration: tests that require scvi-tools and other heavy runtime deps",
]

[tool.black]
line-length = 120

[tool.ruff]
src = ["src"]
line-length = 120
select = [
    "F",  # Errors detected by Pyflakes
    "E",  # Error detected by Pycodestyle
    "W",  # Warning detected by Pycodestyle
    "I",  # isort
    "D",  # pydocstyle
    "B",  # flake8-bugbear
    "TID",  # flake8-tidy-imports
    "C4",  # flake8-comprehensions
    "BLE",  # flake8-blind-except
    "UP",  # pyupgrade
    "RUF100",  # Report unused noqa directives
]
ignore = [
    # line too long -> we accept long comment lines; black gets rid of long code lines
    "E501",
    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
    "E731",
    # allow I, O, l as variable names -> I is the identity matrix
    "E741",
    # Missing docstring in public package
    "D104",
    # Missing docstring in public module
    "D100",
    # Missing docstring in __init__
    "D107",
    # Errors from function calls in argument defaults. These are fine when the result is immutable.
    "B008",
    # __magic__ methods are are often self-explanatory, allow missing docstrings
    "D105",
    # first line should end with a period [Bug: doesn't work with single-line docstrings]
    "D400",
    # First line should be in imperative mood; try rephrasing
    "D401",
    ## Disable one in each pair of mutually incompatible rules
    # We don’t want a blank line before a class docstring
    "D203",
    # We want docstrings to start immediately after the opening triple quote
    "D213",
    "D205"
]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.per-file-ignores]
"docs/*" = ["I"]
"tests/*" = ["D"]
"*/__init__.py" = ["F401"]

[tool.cruft]
skip = [
    "tests",
    "src/**/__init__.py",
    "src/**/basic.py",
    "docs/api.md",
    "docs/changelog.md",
    "docs/references.bib",
    "docs/references.md",
    "docs/notebooks/example.ipynb"
]
